+++ signalworks/tracking/__init__.py
@@ -1,3 +1,10 @@
 # -*- coding: utf-8 -*-
-from .tracking import Event, Track, TimeValue, Partition, MultiChannelError  # noqa: F401
+from .tracking import Track# noqa: F401
+from .error import MultiChannelError, LabreadError
 from .wave import Wave
+from .event import Event
+from .partition import Partition
+from .metatrack import MetaTrack
+from .timevalue import TimeValue
+from .label import Label
+from .value import Value
