+++ signalworks/tracking/tracking.py
@@ -22,14 +22,17 @@ from collections import Iterable
 
 import numpy
 from scipy.io.wavfile import read as wav_read, write as wav_write
-from signalworks.tracking import Wave
+from signalworks.tracking.wave import Wave
+from signalworks.tracking.metatrack import MetaTrack
+from signalworks.tracking.partition import Partition
+from signalworks.tracking.timevalue import TimeValue
 
 logger = logging.getLogger()
 logger.setLevel(logging.DEBUG)
 # logger.setLevel(logging.WARNING)
 # logger.setLevel(logging.ERROR)
 
-__all__ = ["Track", "Event", "Wave", "Label", "Value", "TimeValue", "Partition"]
+__all__ = ["Track", "Label", "Value"]
 
 TIME_TYPE = numpy.int64
 
@@ -83,20 +86,7 @@ def convert_dtype(source, target_dtype):
                 else:  # target_dtype == numpy.int32
                     return (source * (1 << 31)).astype(target_dtype)  # dither?
 
-
-class Error(Exception):
-    pass
-
-
-class LabreadError(Error):
-    pass
-
-
-class MultiChannelError(Error):
-    pass
-
-
-class Track(metaclass=abc.ABCMeta):
+class Track(MetaTrack):
     default_suffix = ".trk"
 
     def __init__(self, path):
@@ -134,9 +124,8 @@ class Track(metaclass=abc.ABCMeta):
 
     fs = property(get_fs, set_fs, doc="sampling frequency")
 
-    @abc.abstractmethod
     def get_duration(self):
-        raise NotImplementedError
+        pass
 
     def set_duration(self, duration):
         raise NotImplementedError
@@ -149,13 +138,11 @@ class Track(metaclass=abc.ABCMeta):
     def __ne__(self, other):
         raise NotImplementedError
 
-    @abc.abstractmethod
     def __len__(self):
-        raise NotImplementedError
+        pass
 
-    @abc.abstractmethod
     def __str__(self):
-        raise NotImplementedError
+        pass
 
     def __add__(self, other):
         raise NotImplementedError
@@ -204,292 +191,6 @@ class Track(metaclass=abc.ABCMeta):
         return t
 
 
-# class Wave(Track):
-#     """monaural waveform"""
-#
-#     default_suffix = ".wav"
-#
-#     def __init__(
-#         self, value: numpy.ndarray, fs, duration=None, offset=0, path=None
-#     ) -> None:
-#         super().__init__(path)
-#         assert isinstance(value, numpy.ndarray)
-#         assert 1 <= value.ndim, "only a single channel is supported"
-#         assert isinstance(fs, int)
-#         assert fs > 0
-#         self._value = value
-#         self._fs = fs
-#         self._offset = (
-#             offset
-#         )  # this is required to support heterogenous fs in multitracks
-#         self.type = "Wave"
-#         self.label = f"amplitude-{value.dtype}"
-#         if not duration:
-#             duration = len(self._value)
-#         assert (
-#             len(self._value) <= duration < len(self._value) + 1
-#         ), "Cannot set duration of a wave to other than a number in " "[length, length+1), where length = len(self.value)"
-#         self._duration = duration
-#
-#     def get_offset(self):
-#         return self._offset
-#
-#     def set_offset(self, offset):
-#         assert 0.0 <= offset < 1.0
-#         self._offset = offset
-#
-#     offset = property(get_offset, set_offset)
-#
-#     def get_time(self):
-#         return (
-#             numpy.arange(len(self._value))
-#             if self._offset == 0
-#             else numpy.arange(len(self._value)) + self._offset
-#         )
-#
-#     def set_time(self, time):
-#         raise Exception("can't set times for Wave")
-#
-#     time = property(get_time, set_time)
-#
-#     def get_value(self):
-#         return self._value
-#
-#     def set_value(self, value):
-#         assert isinstance(value, numpy.ndarray)
-#         assert 1 == value.ndim, "only a single channel is supported"
-#         self._value = value
-#         if not (len(self._value) <= self._duration < len(self._value) + 1):
-#             self._duration = len(self._value)
-#
-#     value = property(get_value, set_value)
-#
-#     def get_duration(self):
-#         return self._duration
-#
-#     def set_duration(self, duration):
-#         assert (
-#             len(self._value) <= duration < len(self._value) + 1
-#         ), "Cannot set duration of a wave to other than a number in [length, length+1) " "- where length = len(self.value)"
-#         self._duration = duration
-#
-#     duration = property(get_duration, set_duration)
-#
-#     #  def get_channels(self):
-#     #      return 1 if self._value.ndim == 1 else self._value.shape[1] # by convention
-#
-#     #  def set_channels(self, *args):
-#     #      raise Exception("Cannot change wave channels - create new wave instead")
-#
-#     #  channels = property(get_channels)  #, set_channels)
-#
-#     def get_dtype(self):
-#         return self._value.dtype
-#
-#     #  def set_dtype(self, *args):
-#     #      raise Exception("Cannot change wave dtype - create new wave instead")
-#     dtype = property(get_dtype)
-#
-#     def get_bitdepth(self):
-#         dtype = self._value.dtype
-#         if dtype == numpy.int16:
-#             return 16
-#         elif dtype == numpy.float32 or dtype == numpy.int32:
-#             return 32
-#         elif dtype == numpy.float64:
-#             return 64
-#         else:
-#             raise Exception("unknown dtype = %s" % dtype)
-#
-#     bitdepth = property(get_bitdepth)
-#
-#     def __eq__(self, other):
-#         if (
-#             (self._fs == other._fs)
-#             and (self._duration == other._duration)
-#             and (len(self._value) == len(other._value))
-#             and (self._value == other._value).all()
-#         ):
-#             return True
-#         else:
-#             return False
-#
-#     def __ne__(self, other):
-#         return not self.__eq__(other)
-#
-#     def __str__(self):
-#         return "\n".join(
-#             [
-#                 f"value={self.value}",
-#                 f"min={self.value.min()}",
-#                 f"max={self.value.max()}",
-#                 f"dtype={self.dtype}",
-#                 f"fs={self.fs}",
-#                 f"duration={self.duration}",
-#             ]
-#         )
-#
-#     def __len__(self):
-#         return len(self._value)
-#
-#     def __add__(self, other):
-#         assert type(other) == type(self), "Cannot add Track objects of different types"
-#         assert self._fs == other._fs, "sampling frequencies must match"
-#         assert self.value.dtype == other.value.dtype, "dtypes must match"
-#         value = numpy.concatenate((self._value, other._value))
-#         return type(self)(value, self.fs)
-#
-#     def resample(self, fs):
-#         """resample to a certain fs"""
-#         assert isinstance(fs, int)
-#         if fs != self._fs:
-#             # do this here, because multirate loading an external lib is problematic for the iOS port
-#             # from pysig import multirate
-#             #  import fractions
-#             #  return type(self)(multirate.resample(self._value, fractions.Fraction(fs, self._fs)), fs)
-#             # return type(self)(multirate.resample(self._value, self._fs, fs), fs)
-#             raise NotImplementedError
-#         else:
-#             return self
-#
-#     def convert_dtype(self, target_dtype):
-#         """returns a new wave with the waveform in the specified target_dtype"""
-#         # TODO: take care of setting new min and max
-#         return type(self)(
-#             convert_dtype(self._value, target_dtype), self._fs, path=self.path
-#         )
-#
-#     def select(self, a, b):
-#         assert a >= 0
-#         assert a < b  # or a <= b?
-#         assert b <= self.duration
-#         # TODO: modify this for float a and b
-#         return type(self)(self._value[a:b], self._fs)
-#
-#     @classmethod
-#     def read(cls, name, *_args, **_kwargs):
-#         """Loads object from name, adding default extension if missing."""
-#         name_wo_ext = os.path.splitext(name)[0]
-#         if name == name_wo_ext:
-#             name += cls.default_suffix
-#         return cls.read_wav(name)
-#
-#     @classmethod
-#     def read_wav(cls, path, channel=None, mmap=False):
-#         """load waveform from file"""
-#         fs, value = wav_read(path, mmap=mmap)
-#         if value.ndim == 1:
-#             if channel is not None and channel != 0:
-#                 raise MultiChannelError(
-#                     "cannot select channel {} from monaural file {}".format(
-#                         channel, path
-#                     )
-#                 )
-#         if value.ndim == 2:
-#             if channel is None:
-#                 raise MultiChannelError(
-#                     f"must select channel when loading file {path} with {value.shape[1]} channels"
-#                 )
-#             try:
-#                 value = value[:, channel]
-#             except IndexError:
-#                 raise MultiChannelError(
-#                     f"cannot select channel {channel} from file "
-#                     f"{path} with {value.shape[1]} channels"
-#                 )
-#         wav = Wave(value, fs, path=path)
-#         if value.dtype == numpy.int16:
-#             wav.min = -32767
-#             wav.max = 32768
-#         else:
-#             raise NotImplementedError
-#         return wav
-#
-#     wav_read = read_wav
-#
-#     def write(self, name, *_args, **_kwargs):
-#         """Saves object to name, adding default extension if missing."""
-#         name_wo_ext = os.path.splitext(name)[0]
-#         if name == name_wo_ext:
-#             name += self.default_suffix
-#         self.write_wav(name)
-#
-#     def write_wav(self, name):
-#         """save waveform to file
-#         The bits-per-sample will be determined by the data-type (mostly)"""
-#         wav_write(name, self._fs, self._value)
-#
-#     wav_write = write_wav
-#
-#     def __getitem__(self, index):
-#         return Wave(self._value[index], self.fs)
-#
-#     # def __setitem__(self, index, value):
-#     #    self._value[index] = value
-#
-#     # def __add__(self, other):
-#     #     """wave1 + wave2"""
-#     #     if self.fs != other.fs:
-#     #         raise Exception("sampling frequency of waves must match")
-#     #     return type(self)(numpy.concatenate((self.va, other.va)), self.fs)  # return correct (child) class
-#
-#     # def delete(self, a, b, fade = 0):
-#     #     pass
-#
-#     # def cut(self, a, b, fade = 0):
-#     #     wave = self.copy(a, b)
-#     #     self.delete(a, b, fade)
-#     #     return wave
-#
-#     # def insert(self, wave, a, fade = 0):
-#     #     """insert wave into self at time a"""
-#     #     if self.fs != wave.fs:
-#     #         raise Exception("sampling frequency of waves must match")
-#     #     if fade:
-#     #         n = round(fade * self.fs)
-#     #         if n*2 > len(wave.signal):
-#     #             raise Exception("fade inverval is too large")
-#     #         up = numpy.linspace(0, 1, n)
-#     #         down = numpy.linspace(1, 0, n)
-#     #         p = wave.signal.copy()
-#     #         p[:n] *= up
-#     #         p[-n:] *= down
-#     #         l = self.signal[:a+n]
-#     #         l[-n:] *= down
-#     #         r = self.signal[a-n:]
-#     #         r[:n] *= up
-#     #     else:
-#     #         self.signal = numpy.concatenate((self.signal[:a], wave.signal, self.signal[a:]))
-#
-#     def crossfade(self, wave, length):
-#         """append wave to self, using a crossfade of a specified length in samples"""
-#         assert type(self) == type(wave), "Cannot add Track objects of different types"
-#         assert self.fs == wave.fs, "sampling frequency of waves must match"
-#         assert isinstance(length, int)
-#         assert length > 0
-#         assert wave.duration >= length
-#         assert self.duration >= length
-#         ramp = numpy.linspace(1, 0, length + 2)[1:-1]  # don't include 0 and 1
-#         value = self.value.copy()
-#         value[-length:] = value[-length:] * ramp + wave.value[:length] * (
-#             1 - ramp
-#         )  # TODO: think about dtypes here
-#         value = numpy.concatenate((value, wave.value[length:]))
-#         return type(self)(value, self.fs)
-#
-#     # TODO: Test / fix me!
-#     def time_warp(self, x, y):
-#         raise NotImplementedError
-#         logger.warning(
-#             "time_warping wave, most of the time this is not what is desired"
-#         )
-#         time = numpy.arange(len(self._value))
-#         # time = index / self._fs
-#         time = numpy.round(numpy.interp(time, x, y)).astype(numpy.int)
-#         # index = int(time * self.fs)
-#         self._value = self._value[time]
-
-
 def get_track_classes() -> List[Track]:
     def all_subclasses(c):
         return c.__subclasses__() + [
@@ -504,1556 +205,6 @@ def get_track_classes() -> List[Track]:
 #  wonder if I can extend Event itself with optional values...
 # class NamedEvent(_Track):
 #  def __init__(self, time, value, fs, duration)
-
-
-class Label(Track):
-    """Like Partition, but label regions do NOT have to be contiguous"""
-
-    default_suffix = ".lab"
-
-    def check(self):
-        """value[k] has beginning and ending times time[2*k] and time[2*k+1]"""
-        assert isinstance(self._time, numpy.ndarray)
-        assert self._time.ndim == 1
-        assert self._time.dtype == TIME_TYPE
-        # assert (numpy.diff(self._time.astype(numpy.float)) >= 0).all(),\
-        # "times must be (non-strictly) monotonically increasing"
-        if not (numpy.diff(self._time.astype(numpy.float)) >= 0).all():
-            logger.warning(
-                "Label times must be (non-strictly) monotonically increasing"
-            )
-        assert isinstance(self._value, numpy.ndarray)
-        # assert self._value.ndim == 1 # TODO: can I remove this?
-        assert isinstance(self._fs, int)
-        assert self._fs > 0
-        assert (
-            numpy.diff(self._time[::2]) > 0
-        ).all(), (
-            "zero-duration labels are not permitted (but abutting labels are permitted)"
-        )
-        # this means an empty partition contains one time value at 0!!!
-        assert len(self._time) == 2 * len(
-            self._value
-        ), "length of time and value *2 must match"
-        assert isinstance(self._duration, TIME_TYPE) or isinstance(self._duration, int)
-        assert not (
-            len(self._time) and self._duration < self._time[-1]
-        ), "duration is not >= times"
-        return True
-
-    def __init__(self, time, value, fs, duration, path=None):
-        super().__init__(path)
-        self._time = time
-        self._value = value
-        self._fs = fs
-        self._duration = duration
-        self.type = "Label"
-        assert self.check()
-
-    def get_time(self):
-        if __debug__:
-            self.check()
-        return self._time
-
-    def set_time(self, time):
-        self._time = time
-        if __debug__:
-            self.check()
-
-    time = property(get_time, set_time)
-
-    def get_value(self):
-        if __debug__:
-            self.check()
-        return self._value
-
-    def set_value(self, value):
-        self._value = value
-        if __debug__:
-            self.check()
-
-    value = property(get_value, set_value)
-
-    def get_duration(self):
-        return self._duration
-
-    def set_duration(self, duration):
-        assert isinstance(duration, (TIME_TYPE, int, numpy.int64))
-        if len(self.time) > 0:
-            assert duration > self.time[-1]
-        self._duration = duration
-
-    duration = property(get_duration, set_duration, doc="duration of track")
-
-    def __str__(self):
-        # s = [u"0"]
-        # for i in range(len(self._value)):
-        #    s.append(':%s:%i/%.3f' % (self._value[i], self._time[i + 1], self._time[i + 1] / self._fs))
-        s = [""]
-        for i in range(len(self._value)):
-            s.append(
-                f"#{i}: %i/%.3f: %s :%i/%.3f\n"
-                % (
-                    self._time[2 * i],
-                    self._time[2 * i] / self._fs,
-                    self._value[i],
-                    self._time[2 * i + 1],
-                    self._time[2 * i + 1] / self._fs,
-                )
-            )
-        s = "".join(s)
-        return "%sfs=%i\nduration=%i" % (s, self._fs, self._duration)
-
-    def __add__(self, other):
-        if self._fs != other._fs:
-            raise Exception("sampling frequencies must match")
-        time = numpy.hstack(
-            (self._time, self.duration + other._time)
-        )  # other._time[0] == 0
-        value = numpy.hstack((self._value, other._value))
-        duration = self.duration + other.duration
-        return Label(time, value, self.fs, duration)
-
-    def __len__(self):
-        return len(self._value)
-
-    def resample(self, fs):
-        """resample to a certain fs"""
-        if fs != self._fs:
-            factor = fs / self._fs
-            time = numpy.round(factor * self._time).astype(TIME_TYPE)
-            time[-1] = numpy.ceil(factor * self._time[-1])
-            assert (
-                numpy.diff(time) > 0
-            ).all(), (
-                "new fs causes times to fold onto themselves due to lack in precision"
-            )
-            duration = numpy.round(factor * self._duration).astype(TIME_TYPE)
-            return type(self)(time, self._value, fs, duration)
-        else:
-            return self
-
-    def select(self, a, b):
-        assert 0 <= a
-        assert a < b
-        assert b <= self.duration
-        if __debug__:  # consistency check
-            self.check()
-        ai = self._time.searchsorted(a)
-        bi = self._time.searchsorted(b)
-        if ai % 2 == 1:  # inside label
-            ai -= 1
-        if bi % 2 == 1:  # inside label
-            bi += 1
-        value = self._value[(ai // 2) : (bi // 2)]
-        time = self.time[ai:bi]
-        time = time - a
-        duration = b - a
-        if time[0] < 0:
-            time[0] = 0
-        if time[-1] > duration:
-            time[-1] = duration
-        return type(self)(time.astype(TIME_TYPE), value, self.fs, duration)
-
-    @classmethod
-    def read_lbl(cls, path, fs=300000, encoding="UTF8"):
-        """load times, values from a .lbl label file"""
-        with open(path, "r", encoding=encoding) as f:
-            lines = f.readlines()
-        if len(lines) == 0:
-            raise ValueError("file '{}' is empty".format(path))
-        lines += "\n"  # make sure it's terminated
-        time = []
-        value = []
-        # label_type = numpy.float64
-        for i, line in enumerate(lines):
-            if line != "\n":
-                try:
-                    t1, t2, label = line.split()
-                except ValueError:
-                    logger.warning(
-                        'ignoring line "%s" in file %s at line %i' % (line, path, i + 1)
-                    )
-                    continue
-                t1 = float(t1)  # / 1000  # this particular
-                t2 = float(t2)  # / 1000  # file format
-                if label[-1] == "\r":
-                    label = label[:-1]
-                dur = t2 - t1
-                if dur > 0:
-                    time.extend([t1, t2])
-                    value.append(label)
-                elif dur == 0:
-                    logger.warning(
-                        'zero-length label "%s" in file %s at line %i, ignoring'
-                        % (label, path, i + 1)
-                    )
-                else:
-                    raise LabreadError(
-                        "label file contains times that are not monotonically increasing"
-                    )
-        if len(time) == 0:
-            raise (Exception("file is empty or all lines were ignored"))
-        time = numpy.round(numpy.array(time) * fs).astype(TIME_TYPE)
-        # assert labels are not longer than X characters
-        value = numpy.array(value)
-        # best guess at duration (+1 sec)
-        lab = Label(time, value, fs=fs, duration=TIME_TYPE(time[-1] + fs), path=path)
-        return lab
-
-    @classmethod
-    def read(cls, *args, **kwargs):
-        return cls.read_lbl(cls, *args, **kwargs)
-
-
-class Value(Track):
-    """can store a singular value (e.g. comment string) of any type"""
-
-    def __init__(self, value, fs, duration, path):
-        super().__init__(path)
-        assert isinstance(fs, int)
-        assert fs > 0
-        assert isinstance(duration, TIME_TYPE) or isinstance(duration, int)
-        self._value = value
-        self._fs = fs
-        self._duration = duration
-
-    def __getitem__(self, k):
-        return self.value[k]
-
-    def __setitem__(self, k, v):
-        self.value[k] = v
-
-    def __eq__(self, other):
-        if (
-            (self._fs == other._fs)
-            and (self._duration == other._duration)
-            and isinstance(other.value, type(self.value))
-        ):
-            return self.value == other.value
-        else:
-            return False
-
-    def __ne__(self, other):
-        return not self.__eq__(other)
-
-    def __len__(self):
-        return 1
-
-    def __str__(self):
-        return self._value
-
-    def get_duration(self):
-        return self._duration
-
-    def set_duration(self, duration):
-        assert isinstance(duration, TIME_TYPE) or isinstance(duration, int)
-        self._duration = duration
-
-    duration = property(get_duration, set_duration, doc="duration of track")
-
-    def get_value(self):
-        return self._value
-
-    def set_value(self, value):
-        self._value = value
-
-    value = property(get_value, set_value)
-
-    def select(self, a, b):
-        raise NotImplementedError
-
-    def time_warp(self, _x, y):
-        self.duration = y[-1]  # changed this from _duration
-
-    @classmethod
-    def identify_ext(cls, name, track_name=None):
-        if name == os.path.splitext(name)[0]:
-            if track_name is None:
-                raise ValueError("neither track name nor extension specified")
-            elif track_name == "input":
-                ext = ".txt"
-            elif track_name == "xml":
-                ext = ".xml"
-            elif track_name == "search":
-                ext = ".search"
-            elif track_name == "pronunciation":
-                ext = ".pron"
-            else:
-                raise ValueError("unknown track name '{}'".format(track_name))
-        else:
-            ext = os.path.splitext(name)[1]
-        return ext
-
-    @classmethod
-    def read(cls, name, fs, duration, track_name=None, *_args, **_kwargs):
-        """Loads object from name, adding extension if missing."""
-        ext = cls.identify_ext(name, track_name)
-        name = os.path.splitext(name)[0] + ext
-        if ext in (".txt", ".xml", ".search"):
-            with open(name, "r") as f:
-                value = f.read()
-                if ext == ".search":
-                    value = str(value, "UTF-8")
-                self = Value(value, fs, duration)
-        elif ext == ".pron":
-            with open(name, "r") as f:
-                pron = [tuple(line.split("\t")) for line in f.read().split("\n")]
-                self = Value(pron, fs, duration)
-        else:
-            raise ValueError("unknown extension '{}'".format(ext))
-        return self
-
-    def write(self, name, track_name=None, *_args, **_kwargs):
-        """
-        Saves object to name, adding extension if missing.
-        """
-        ext = self.identify_ext(name, track_name)
-        name = os.path.splitext(name)[0] + ext
-        if ext in (".txt", ".xml", ".search"):
-            with open(name, "w") as f:
-                f.write(self.get_value())
-        elif ext == ".pron":
-            with open(name, "w") as f:
-                f.write(
-                    "\n".join(word + "\t" + pron for word, pron in self.get_value())
-                )
-        else:
-            raise ValueError("unknown extension '{}'".format(ext))
-
-
-class TimeValue(Track):
-    def __init__(self, time, value, fs, duration, path=None):
-        super().__init__(path)
-        assert isinstance(time, numpy.ndarray)
-        assert time.ndim == 1
-        assert time.dtype == TIME_TYPE
-        assert (
-            numpy.diff(time.astype(numpy.float)) > 0
-        ).all(), "times must be strictly monotonically increasing"
-        assert isinstance(value, numpy.ndarray)
-        assert isinstance(fs, int)
-        assert fs > 0
-        assert isinstance(duration, TIME_TYPE) or isinstance(duration, int)
-        assert len(time) == len(value), "length of time and value must match"
-        assert not (len(time) and duration <= time[-1]), "duration is not > times"
-        self._time = time
-        self._value: numpy.ndarray = value
-        self._fs = fs
-        self._duration = duration
-        self.min = numpy.nanmin(value)
-        self.max = numpy.nanmax(value)
-        self.unit = ""
-        self.label = ""
-        self.path = path
-
-    def get_time(self):
-        assert (
-            numpy.diff(self._time.astype(numpy.float)) > 0
-        ).all(), (
-            "times must be strictly monotonically increasing"
-        )  # in case the user messed with .time[index] directly
-        return self._time
-
-    def set_time(self, time):
-        assert isinstance(time, numpy.ndarray)
-        assert time.ndim == 1
-        assert time.dtype == TIME_TYPE
-        assert not (len(time) and self._duration <= time[-1]), "duration is not > times"
-        assert (
-            numpy.diff(time.astype(numpy.float)) > 0
-        ).all(), "times must be strictly monotonically increasing"
-        assert len(time) == len(self._value), "length of time and value must match"
-        self._time = time
-
-    time = property(get_time, set_time)
-
-    def get_value(self):
-        return self._value
-
-    def set_value(self, value):
-        assert isinstance(value, numpy.ndarray)
-        assert len(self._time) == len(value), "length of time and value must match"
-        self._value = value
-
-    value = property(get_value, set_value)
-
-    def get_duration(self):
-        return self._duration
-
-    def set_duration(
-        self, duration
-    ):  # assume times are available, if not, this must be overridden
-        assert isinstance(duration, TIME_TYPE) or isinstance(duration, int)
-        assert not (
-            len(self._time) and duration <= self._time[-1]
-        ), "duration is not > times"
-        self._duration = duration
-
-    duration = property(get_duration, set_duration, doc="duration")
-
-    def __eq__(self, other):
-        if (
-            (self._fs == other._fs)
-            and (self._duration == other._duration)
-            and (len(self._time) == len(other._time))
-            and (len(self._value) == len(other._value))
-            and (self._time == other._time).all()
-            and numpy.allclose(
-                numpy.round(self._value, 3), numpy.round(other._value, 3)
-            )
-        ):
-            return True
-        else:
-            return False
-
-    def __ne__(self, other):
-        return not self.__eq__(other)
-
-    def __len__(self):
-        return len(self._time)
-
-    def __str__(self):
-        return f"{list(zip(self._time, self._value))}, fs={self._fs}, duration={self._duration}, path={self.path}"
-
-    def __add__(self, other):
-        assert type(other) == type(self), "Cannot add Track objects of different types"
-        assert self.fs == other.fs, "sampling frequencies must match"
-        time = numpy.concatenate(
-            (self.time, (other.time + self.duration).astype(other.time.dtype))
-        )
-        value = numpy.concatenate((self.value, other.value))
-        duration = self.duration + other.duration
-        return type(self)(time, value, self.fs, duration)
-
-    # def __iadd__(self, other):
-    #     assert type(other) == type(self), "Cannot add Track objects of different types"
-    #     assert self.fs == other.fs, "sampling frequencies must match"
-    #     self._time = numpy.concatenate((self._time, (other._time + self._duration).astype(other.time.dtype)))
-    #     self._value = numpy.concatenate((self._value, other.value))
-    #     duration = self.duration + other.duration
-    #     return self
-
-    def resample(self, fs):
-        if fs != self._fs:
-            factor = fs / self._fs
-            time = numpy.round(factor * self._time).astype(TIME_TYPE)
-            assert (
-                numpy.diff(time) > 0
-            ).all(), (
-                "new fs causes times to fold onto themselves due to lack in precision"
-            )
-            # need to use numpy.round for consistency - it's different from the built-in round
-            duration = int(numpy.ceil(factor * self._duration))
-            return type(self)(time, self._value, fs, duration)
-        else:
-            return self
-
-    def select(self, a, b):
-        """(SHOULD!) return a copy"""
-        # assert isinstance(a, TIME_TYPE)  # this doesn't seem necessary
-        # assert isinstance(b, TIME_TYPE)
-        assert b > a
-        ai = self.time.searchsorted(a)
-        bi = self.time.searchsorted(b)
-        time = self._time[ai:bi] - a  # not possible to make this a view
-        value = self._value[ai:bi]  # this is a view - should probably be a copy.
-        return type(self)(time, value, self.fs, TIME_TYPE(b - a))
-
-    default_suffix = ".tmv"
-
-    @classmethod
-    def read(cls, name, track_name=None, fs=300_000, duration=None, *_args, **_kwargs):
-        """Loads object from name, adding default extension if missing."""
-        if name == os.path.splitext(name)[0]:
-            if track_name:
-                ext = "." + track_name
-            else:
-                ext = cls.default_suffix
-            name += ext
-        else:
-            ext = os.path.splitext(name)[1].lower()
-        if ext in (".pit", ".nrg", cls.default_suffix):
-            self = cls.read_tmv(name, fs)
-        elif ext == ".f0":
-            self = cls.read_pitchtier(name, fs)[0]
-        elif ext == ".pitchtier":
-            self = cls.read_pitchtier(name, fs)
-        else:
-            self = None
-            try:
-                if self is None:
-                    self = cls.read_f0(name, fs)[0]
-            except FileNotFoundError:  # TODO: need more exceptions?
-                pass
-            try:
-                if self is None:
-                    self = cls.read_pitchtier(name, fs)
-            except FileNotFoundError:  # TODO: need more exceptions?
-                pass
-            if self is None:
-                raise ValueError("file '{}' has unknown format".format(name))
-        if duration:
-            self.set_duration(duration)
-        self.path = name
-        return self
-
-    @classmethod
-    def read_f0(cls, name, frameRate=0.01, frameSize=0.0075, fs=48000):
-        # return TimeValue(numpy.ndarray([1]).astype(TIME_TYPE), numpy.ndarray([1]), fs, 1)
-        # 4 fields for each frame, pitch, probability of voicing, local root mean squared measurements,
-        # and the peak normalized cross-correlation value frame arguments are in seconds
-        f = open(name, "r")
-        lines = f.readlines()
-        f.close()
-        F = len(lines)
-        time = numpy.round((numpy.arange(F) * frameRate + frameSize / 2) * fs).astype(
-            TIME_TYPE
-        )
-        duration = numpy.round((F * frameRate + frameSize) * fs).astype(TIME_TYPE)
-        # time = numpy.arange(len(lines)) * frameRate + frameSize / 2
-        value_f0 = numpy.zeros(F, numpy.float32)
-        value_vox = numpy.zeros(F, numpy.float32)
-        import re
-
-        form = re.compile(r"^(\S+)\s+(\S+)\s+(\S+)\s+(\S+)$")
-        for i, line in enumerate(lines):
-            match = form.search(line)
-            if not match:
-                # logger.error('badly formatted f0 file: {}'.format("".join(lines)))
-                raise Exception
-                # return (TimeValue(numpy.array([0]), numpy.array([100]), fs, duration),
-                #         Partition(numpy.array([0, duration]), numpy.array([0]), fs))
-            f0, voicing, _energy, _xcorr = match.groups()
-            value_f0[i] = float(f0)
-            value_vox[i] = float(voicing)
-        index = numpy.where(value_f0 > 0)[0]  # keep only nonzeros F0 values
-        pit = TimeValue(time[index], value_f0[index], fs, duration)
-        vox = Partition.from_TimeValue(
-            TimeValue(time, value_vox, fs, duration)
-        )  # return a Partition
-        return pit, vox
-
-    f0read = read_f0
-
-    @classmethod
-    def read_tmv(cls, name, fs=300000):
-        obj = numpy.loadtxt(name)
-        time = obj[:, 0]
-        value = obj[:, 1]
-        time = numpy.round(time * fs).astype(TIME_TYPE)
-        duration = (time[-1] + 1).astype(TIME_TYPE)
-        return TimeValue(time, value, fs, duration, path=name)
-
-    @classmethod
-    def read_frm(cls, name, fs=48000):
-        f = open(name, "r")
-        lines = f.readlines()
-        f.close()
-        frameRate = 0.01
-        frameSize = 0.049
-        time = numpy.round(
-            (numpy.arange(len(lines)) * frameRate + frameSize / 2) * fs
-        ).astype(TIME_TYPE)
-        duration = (time[-1] + 1).astype(TIME_TYPE)
-        for i, line in enumerate(lines):
-            fb = numpy.array(line.split()).astype(numpy.float)
-            if i == 0:
-                value = numpy.zeros((len(lines), len(fb)), numpy.float64)
-            value[i, :] = fb
-        return TimeValue(time, value, fs, duration, path=name)
-
-    frmread = read_frm
-
-    def write(self, name, track_name=None, *_args, **_kwargs):
-        """Saves object to name, adding default extension if missing."""
-        name_wo_ext = os.path.splitext(name)[0]
-        if name == name_wo_ext:
-            if track_name:
-                ext = "." + track_name
-            else:
-                ext = self.default_suffix
-            name += ext
-        self.write_tmv(name)
-
-    def write_f0(self, name, frameRate=0.01, frameSize=0.0075, fs=48000):
-        raise NotImplementedError("TimeValue.write_f0: not yet implemented")
-
-    def write_tmv(self, name):
-        obj = numpy.c_[self._time / self._fs, self._value]
-        numpy.savetxt(name, obj)
-        # with open(name, 'w') as f:
-        #     for tv in zip(self._time / self._fs, self._value):
-        #         f.write('\t'.join(str(round(x, 16)) for x in tv) + '\n')
-
-    def write_frm(self, name):
-        with open(name, "w") as f:
-            if isinstance(self._value[0], Iterable):
-                for fb in self._value:
-                    f.write("\t".join(str(round(f, 3)) for f in fb) + "\n")
-            else:
-                for fb in self._value:
-                    f.write(str(round(fb, 3)) + "\n")
-
-    @classmethod
-    def from_Partition(cls, p):
-        """convert a partition track into a time-value track"""
-        assert isinstance(p, Partition)
-        time = numpy.r_[p.time[:-1], p.time[-1] - 1, p.time[1:-1] - 1]
-        time.sort()
-        # double this to keep it constant
-        value = numpy.array(
-            [p.value[int(i)] for i in numpy.arange(0, len(p.value), 0.5)]
-        )
-        return TimeValue(time, value, p.fs, p.duration)
-
-    # def __getitem__(self, index):  # should I make the default just the value?
-    #     return (self._time[index], self._value[index])
-    #     return self._value[index] # ??
-
-    # def __setitem__(self, index, value):
-    #     self._time[index] = value[0]
-    #     self._value[index] = value[1]
-
-    def get_index(self, t):
-        """return the index of the nearest available data"""
-        if len(self._time) == 0:
-            raise Exception
-        if t <= 0:
-            return 0
-        if t >= self._time[-1]:
-            return len(self._value) - 1
-        # t is within bounds
-        ri = self._time.searchsorted(t)
-        li = ri - 1
-        rt = self._time[ri]
-        lt = self._time[li]
-        if (t - lt) < (rt - t):
-            return li
-        else:
-            return ri
-
-    def _get_value(self, t, interpolation="nearest"):
-        # check bounds
-        if len(self._time) == 0:
-            raise Exception
-        if t < self._time[0]:
-            # raise Exception('out of bounds left')
-            # logger.warning('out of bounds left')
-            v = self._value[0]  # better? yes!
-        elif t > self._time[-1]:
-            # raise Exception('out of bounds right')
-            # logger.warning('out of bounds right')
-            v = self._value[-1]  # better? yes!
-        else:  # t is within bounds
-            ri = self._time.searchsorted(t)  # a numpy function
-            rt = self._time[ri]
-            rv = self._value[ri]
-            if rt == t:  # exact hit
-                v = rv
-            else:
-                li = ri - 1
-                lt = self._time[li]
-                lv = self._value[li]
-                if interpolation == "nearest":
-                    if (t - lt) < (rt - t):
-                        v = lv
-                    else:
-                        v = rv
-                else:  # linear
-                    a = float(t - lt) / float(rt - lt)
-                    v = a * rv + (1.0 - a) * lv
-        return v
-
-    def get(
-        self, T, interpolation="nearest"
-    ):  # start using interp_* methods, which are faster
-        """return the values at times T"""
-        if isinstance(T, numpy.ndarray):
-            assert T.ndim == 1
-            n = (
-                self._value.shape[1] if self._value.ndim > 1 else 1
-            )  # this can be done much better I'm sure
-            V = numpy.empty((T.shape[0], n), dtype=self._value.dtype)
-            for i, t in enumerate(T):
-                # TODO: speed this up by using a cached version of the interpolate function
-                V[i] = self._get_value(t, interpolation)
-            return V
-        else:
-            return self._get_value(T, interpolation)
-
-    def interp_linear(self, T):
-        assert numpy.all(numpy.diff(self.time) > 0)
-        return numpy.interp(T, self.time, self.value)
-
-    def interp_nearest(self, T):
-        assert numpy.all(numpy.diff(self.time) > 0)
-        index = numpy.interp(T, self.time, numpy.arange(T))
-        return self.value[numpy.round(index)]  # TODO: test me!
-
-    def interpolate(self, T, kind="linear"):
-        """kind : str or int, optional
-        Specifies the kind of interpolation as a string
-        ('linear', 'nearest', 'zero', 'slinear', 'quadratic, 'cubic'
-        where 'slinear', 'quadratic' and 'cubic' refer to a spline
-        interpolation of first, second or third order) or as an integer
-        specifying the order of the spline interpolator to use.
-        Default is 'linear'."""
-        from scipy.interpolate import interp1d
-
-        if len(self.time):
-            f = interp1d(
-                self.time,
-                self.value,
-                kind=kind,
-                axis=0,
-                copy=True,
-                bounds_error=True,
-                fill_value=numpy.nan,
-            )
-            return f(T)
-        else:
-            logger.warning("interpolating without data")
-            return numpy.ones(len(T)) * numpy.nan
-
-    # def select_index(self, a, b):
-    #     """return indeces such that a <= time < b"""
-    #     assert b > a
-    #     return range(self.time.searchsorted(a), self.time.searchsorted(b))
-
-    # def select(self, a, b):
-    #     """copy a section of the track, interval [)"""
-    #     #indexa = numpy.where((self.ti - a) >= 0)[0][0] # first inside
-    #     #indexb = numpy.where((self.ti - b) <= 0)[0][-1] # last inside
-    #     index = self.select_index(a, b)
-    #     if len(index) == 0:
-    #         if type(self._value) == numpy.ndarray:
-    #             value = numpy.array([])
-    #         else:
-    #             value = []
-    #         return type(self)(time=numpy.empty(0, dtype=TIME_TYPE), value=value, fs=self.fs, duration=b-a)
-    #     # on the line above, I first had [0], which cause all subsequent += to be in int, causing a bad bug.
-    #     # Perhaps this is the reason to choose all times as int
-    #     ai = index[0]
-    #     bi = index[-1]
-    #     time = self._time[ai:bi] #.copy() - don't think that's needed here ...
-    #     value = self._value[ai:bi] #.copy()
-    #     ## now, to copy the track accurately, we have to go all the way to the edges, and capture the values there,
-    #     ## in accordance to the underlying interpolation function
-    #     ### limit selection
-    #     ##if a < self._time[0]:
-    #         ##a = self.ti[0]
-    #     ##if b > self.ti[-1]:
-    #         ##b = self.ti[-1]
-    #     ###if a < self._time[index[0]]:  # prepend - not sure if this is needed
-    #         ###time  = numpy.concatenate(([a], time))
-    #         ###value = numpy.concatenate(([self.get(a, interpolation)], value))
-    #     ## can't append in any way that make sense, I _think_ TODO: Check this
-    #     ##if self.time[index[-1]] <= b:
-    #         ##time = numpy.concatenate((time, [b]))
-    #         ##value = numpy.concatenate((value, f(b).T))
-    #     ##time -= time[0]
-    #     time -= a # TODO: Check this!
-    #     tv = type(self)(time=time, value=value, fs=self.fs, duration=b-a)
-    #     assert type(self._value) == type(tv.value)
-    #     return tv
-
-    def time_warp(self, X, Y):
-        assert X[0] == 0
-        assert Y[0] == 0
-        assert X[-1] == self.duration
-        time = numpy.interp(self.time, X, Y).astype(self.time.dtype)
-        if 0:
-            from matplotlib import pylab
-
-            pylab.plot(X, Y, "rx-")
-            for x, y in zip(self.time, time):
-                pylab.plot([x, x, 0], [0, y, y])
-            pylab.show()
-        assert len(numpy.where(numpy.diff(time) <= 0)[0]) == 0  # no collapsed items
-        self._time = time  # don't do duration check yet
-        self.duration = Y[-1]  # changed this from _duration
-
-    def interp(self, time):  # TODO: unify with get()
-        if self._value.ndim == 1:
-            return numpy.interp(
-                time, self._time, self._value, self._value[0], self._value[-1]
-            )
-        elif self._value.ndim == 2:
-            value = numpy.empty(
-                (len(time), self._value.shape[1]), dtype=self._value.dtype
-            )
-            for j in range(self._value.shape[1]):
-                value[:, j] = numpy.interp(
-                    time,
-                    self._time,
-                    self._value[:, j],
-                    self._value[0, j],
-                    self._value[-1, j],
-                )
-            return value
-        else:
-            raise Exception
-
-
-class Partition(Track):
-    default_suffix = ".lab"
-
-    def check(self):
-        assert isinstance(self._time, numpy.ndarray)
-        assert self._time.ndim == 1
-        assert self._time.dtype == TIME_TYPE
-        # assert (numpy.diff(self._time.astype(numpy.float)) > 0).all(),\
-        #     "times must be strictly monotonically increasing"
-        if not (numpy.diff(self._time.astype(numpy.float)) > 0).all():
-            logger.warning("Partition: times must be strictly monotonically increasing")
-        assert isinstance(self._value, numpy.ndarray)
-        # assert self._value.ndim == 1 # TODO: can I remove this?
-        assert isinstance(self._fs, int)
-        assert self._fs > 0
-        # if len(self._time):
-        assert self._time[0] == 0, "partition must begin at time 0"
-        # assert (numpy.diff(self._time) > 0).all(), "zero-duration labels are not permitted"
-        if not (numpy.diff(self._time) > 0).all():
-            logger.warning("Partition: zero-duration labels are not permitted")
-        # this means an empty partition contains one time value at 0!!!
-        assert (
-            len(self._time) == len(self._value) + 1
-        ), "length of time and value+1 must match"
-        # else:
-        #    assert len(self._value) == 0
-        return True
-
-    def __init__(self, time, value, fs, path=None):
-        super().__init__(path)
-        self._time = time
-        self._value = value
-        self._fs = fs
-        assert self.check()
-
-    def get_time(self):
-        # assert (numpy.diff(self._time.astype(numpy.float)) > 0).all(),\
-        # "times must be strictly monotonically increasing" # in case the user messed with .time[index] directly
-        if not (numpy.diff(self._time.astype(numpy.float)) > 0).all():
-            logger.warning("get_time times must be strictly monotonically increasing")
-        return self._time
-
-    def set_time(self, time):
-        assert isinstance(time, numpy.ndarray)
-        assert time.ndim == 1
-        assert time.dtype == TIME_TYPE
-        if len(time):
-            assert time[0] == 0, "partition must begin at time 0"
-            # assert (numpy.diff(time) > 0).all(), "zero-duration labels are not permitted"
-            # assert (numpy.diff(time) >= 0).all(), "zero-duration labels are not permitted"
-            if not (numpy.diff(time) > 0).all():
-                logger.warning("encorter zero-duration labels are not permitted")
-            assert (
-                len(time) == len(self._value) + 1
-            ), "length of time and value+1 must match"
-        else:
-            assert len(self._value) == 0
-        self._time = time
-
-    time = property(get_time, set_time)
-
-    def get_value(self):
-        return self._value
-
-    def set_value(self, value):
-        assert isinstance(value, numpy.ndarray)
-        # assert value.ndim == 1
-        assert (
-            len(self._time) == len(self._value) + 1
-        ), "length of time and value must match"
-        self._value = value
-
-    value = property(get_value, set_value)
-
-    def get_duration(self):
-        if len(self._value):
-            return self._time[-1]
-        else:
-            return 0  # or None, or raise Exception
-
-    def set_duration(self, duration):
-        assert isinstance(duration, TIME_TYPE) or isinstance(duration, int)
-        if len(self._value):
-            assert (
-                duration > self._time[-2]
-            ), "can't set duration to a smaller or equal value than the next-to-last boundary (this would result in losing the last value)"
-            self._time[-1] = duration
-        else:
-            if duration != 0:
-                raise Exception(
-                    "cannot set duration of an empty Partition to anything but 0"
-                )
-
-    duration = property(get_duration, set_duration, doc="duration of track in fs")
-
-    def __eq__(self, other):
-        try:
-            if (
-                (self._fs == other._fs)
-                and (len(self._time) == len(other._time))
-                and (len(self._value) == len(other._value))
-                and (self._time == other._time).all()
-                and (self._value == other._value).all()
-            ):
-                return True
-        except AttributeError:
-            pass
-        return False
-
-    def __ne__(self, other):
-        return not self.__eq__(other)
-
-    def __len__(self):
-        return len(self._value)
-
-    # def resample(self, fs):
-    #     if fs != self._fs:
-    #         factor = fs / self._fs
-    #         self._time = numpy.round(factor * self._time).astype(TIME_TYPE)
-    #         assert (numpy.diff(self._time) > 0).all(),\
-    #             "new fs causes times to fold onto themselves due to lack in precision"
-    #         self._fs = fs
-
-    def resample(self, fs):
-        """resample to a certain fs"""
-        if fs != self._fs:
-            factor = fs / self._fs
-            time = numpy.round(factor * self._time).astype(TIME_TYPE)
-            time[-1] = numpy.ceil(factor * self._time[-1])
-            assert (
-                numpy.diff(time) > 0
-            ).all(), (
-                "new fs causes times to fold onto themselves due to lack in precision"
-            )
-            return type(self)(time, self._value, fs)
-        else:
-            return self
-
-    def select(self, a, b):
-        assert 0 <= a
-        assert a < b
-        assert b <= self._time[-1]
-        if __debug__:  # consistency check
-            self.check()
-        ai = self._time.searchsorted(a)
-        bi = self._time.searchsorted(b)
-        if a < self._time[ai]:  # prepend
-            ai -= 1
-        if self._time[bi] < b:  # append
-            bi += 1
-        value = self._value[ai:bi]
-        time = self.time[
-            ai : bi + 1
-        ].copy()  # otherwise we are going to modify the original below!
-        # if a < self._time[ai]:  # prepend
-        time[0] = a
-        # if self._time[bi] < b:  # append
-        time[-1] = b
-        time = time - a  # must make a copy!
-        return type(self)(time.astype(TIME_TYPE), value, self.fs)
-
-    @classmethod
-    def read(cls, name, fs=300_000, track_name=None, *_args, **_kwargs):
-        """Loads object from name, adding default extension if missing.
-        It figures out which read function is appropriate based on extension.
-        """
-        if name == os.path.splitext(name)[0]:
-            assert track_name, "neither track name nor extension specified"
-            ext = "." + track_name
-            name += ext
-        else:
-            ext = os.path.splitext(name)[1].lower()[1:]
-        if ext in ("lab", "dem", "rec"):
-            return cls.read_lab(name, fs)
-        elif ext == "textgrid":
-            return cls.read_textgrid(name, fs)
-        else:
-            try:
-                return cls.read_lab(name, fs)
-            except FileNotFoundError:  # TODO: need to catch more exceptions?
-                pass
-            try:
-                return cls.read_textgrid(name, fs)
-            except FileNotFoundError:  # TODO: need to catch more exceptions?
-                pass
-            raise ValueError("file '{}' has unknown format".format(name))
-
-    @classmethod
-    def read_lab(cls, name, fs=300_000, encoding="UTF8"):
-        """load times, values from a label file"""
-        with codecs.open(name, "r", encoding=encoding) as f:
-            lines = f.readlines()
-        if len(lines) == 0:
-            raise ValueError("file '{}' is empty".format(name))
-        time = []
-        value = []
-        # label_type = numpy.float64
-        for i, line in enumerate(lines):
-            try:
-                t1, t2, label = line[:-1].split()
-            except ValueError:
-                logger.warning(
-                    'ignoring line "%s" in file %s at line %i' % (line, name, i + 1)
-                )
-                continue
-            t1 = float(t1)
-            t2 = float(t2)
-            if label[-1] == "\r":
-                label = label[:-1]
-            if len(time) == 0:
-                time.append(t1)
-            else:
-                if time[-1] != t1:
-                    logger.warning(
-                        'noncontiguous label "%s" in file %s at line %i, fixing'
-                        % (label, name, i + 1)
-                    )
-            dur = t2 - time[-1]
-            if dur > 0:
-                time.append(t2)
-                value.append(label)
-            elif dur == 0:
-                logger.warning(
-                    'zero-length label "%s" in file %s at line %i, ignoring'
-                    % (label, name, i + 1)
-                )
-            else:
-                raise LabreadError(
-                    "label file contains times that are not monotonically increasing"
-                )
-        if len(time) == 0:
-            raise (Exception("file is empty or all lines were ignored"))
-        if time[0] != 0:
-            logger.warning("moving first label boundary to zero")
-            time[0] = 0
-            # or insert a first label
-            # time.insert(0, 0)
-            # value.insert(0, default_label)
-        time = numpy.round(numpy.array(time) * fs).astype(TIME_TYPE)
-        # assert labels are not longer than 8 characters
-        # value = numpy.array(value, dtype='U16' if label_type is str else numpy.float64)
-        value = numpy.array(
-            value, dtype="U16"
-        )  # if label_type is str else numpy.float64)
-        return Partition(
-            time, value, fs=fs, path=name
-        )  # u1p to 16 characters (labels could be words)
-
-    @classmethod
-    def read_partition(cls, name, fs=48000, encoding="UTF8"):
-        """load times, values from a label file"""
-        with codecs.open(name, "r", encoding=encoding) as f:
-            lines = f.readlines()
-        if len(lines) == 0:
-            raise ValueError("file '{}' is empty".format(name))
-        time = [0]
-        value = []
-        label_type = numpy.float64
-        for i, line in enumerate(lines):
-            try:
-                t, _t, label = line[:-1].split()
-            except ValueError:
-                logger.warning(
-                    'ignoring line "%s" in file %s at line %i' % (line, name, i + 1)
-                )
-                continue
-            t = float(t)
-            if label[-1] == "\r":
-                label = label[:-1]
-            try:
-                label = label_type(label)
-            except ValueError:
-                if len(time) == 1:
-                    label_type = str
-                    label = label_type(label)
-                else:
-                    raise
-            if len(time) == 1:
-                dur = t
-            else:
-                dur = t - time[-1]
-
-            time.append(t)
-            if dur > 0:
-                value.append(label)
-            elif dur == 0:
-                logger.warning(
-                    'zero-length label "%s" in file %s at line %i, ignoring'
-                    % (label, name, i + 1)
-                )
-            else:
-                raise LabreadError(
-                    "label file contains times that are not monotonically increasing"
-                )
-        if len(time) == 1:
-            raise (Exception("file is empty or all lines were ignored"))
-        if time[0] != 0:
-            logger.warning("moving first label boundary to zero")
-            time[0] = 0
-            # or insert a first label
-            # time.insert(0, 0)
-            # value.insert(0, default_label)
-        time = numpy.round(numpy.array(time) * fs).astype(TIME_TYPE)
-        # assert labels are not longer than 8 characters
-        value = numpy.array(value, dtype="U16" if label_type is str else numpy.float64)
-        return Partition(
-            time, value, fs=fs
-        )  # u1p to 16 characters (labels could be words)
-
-    def write(self, name):
-        """Saves object to name, adding default extension if missing."""
-        name_wo_ext = os.path.splitext(name)[0]
-        if name == name_wo_ext:
-            assert name, "neither track name nor extension specified"
-            name += self.default_suffix
-        self.write_lab(name)
-
-    def write_lab(self, file):
-        if hasattr(file, "read"):
-            f = file
-        else:
-            f = open(file, "w")
-        for i, v in enumerate(self._value):
-            f.write(
-                "%f %f %s\n" % (self._time[i] / self.fs, self._time[i + 1] / self.fs, v)
-            )
-        f.close()
-
-    def write_partition(self, name):
-        # written to generate a format exactly like
-        # CMU-ARCTIC lab/*.lab files
-        f = open(name, "w")
-        f.write("#\n")
-        for i, v in enumerate(self._value):
-            f.write("%f 125 %s\n" % (self.fs, self._time[i + 1] / self.fs, v))
-        f.close()
-
-    @classmethod
-    def from_TimeValue(cls, tv):
-        """convert a time value track with repeating values into a partition track"""
-        assert isinstance(tv, TimeValue)
-        boundary = numpy.where(numpy.diff(tv.value))[0]
-        time = numpy.empty(len(boundary), dtype=tv.time.dtype)
-        value = numpy.empty(len(boundary) + 1, dtype=tv.value.dtype)
-        duration = tv.duration
-        for i in range(len(time)):
-            index = boundary[i]
-            time[i] = (tv.time[index] + tv.time[index + 1]) / 2
-        time = numpy.concatenate(([0], time, [duration])).astype(TIME_TYPE)
-        value[0] = tv.value[0]
-        for i in range(len(boundary)):
-            value[i + 1] = tv.value[boundary[i] + 1]
-        par = Partition(time, value, tv.fs, path=tv.path)
-        assert par.check()
-        return par
-
-    def __getitem__(self, index):
-        return self._time[index], self._value[index], self._time[index + 1]
-
-    def __setitem__(self, index, value):
-        raise NotImplementedError  # this would be some work to support
-
-    def __str__(self):
-        s = [""]
-        for i in range(len(self._value)):
-            s.append(
-                "#%i: %i/%.3f: %s :%i/%.3f\n"
-                % (
-                    i,
-                    self._time[i],
-                    self._time[i] / self._fs,
-                    self._value[i],
-                    self._time[i + 1],
-                    self._time[i + 1] / self._fs,
-                )
-            )
-        s = "".join(s)
-        return "%s\nfs=%i\nduration=%i" % (s, self._fs, self.duration)
-
-    def __add__(self, other):
-        if self._fs != other._fs:
-            raise Exception("sampling frequencies must match")
-        time = numpy.hstack(
-            (self._time, self._time[-1] + other._time[1:])
-        )  # other._time[0] == 0
-        value = numpy.hstack((self._value, other._value))
-        # duration = self.duration + other.duration
-        return Partition(time, value, self.fs)
-
-    def crossfade(self, partition, length):
-        """append to self, using a crossfade of a specified length in samples"""
-        assert type(self) == type(
-            partition
-        ), "Cannot add Track objects of different types"
-        assert self.fs == partition.fs, "sampling frequency of waves must match"
-        assert isinstance(length, int)
-        assert length > 0
-        assert partition.duration >= length
-        assert self.duration >= length
-        # cut left, cut right, concatenate, could be sped up
-        a = self.select(0, self.duration - length // 2)
-        b = partition.select(length - length // 2, partition.duration)
-        return a + b
-
-    def get(self, t):
-        """returns current label at time t"""
-        # return self.value[numpy.where((self.time - t) <= 0)[0][-1]] # last one that is <= 0
-        return self._value[
-            (self._time.searchsorted(t + 1) - 1).clip(0, len(self._value) - 1)
-        ]
-
-    def append(self, time, value):
-        """appends a value at the end, time is new endtime"""
-        assert time > self._time[-1]
-        self._time = numpy.hstack((self._time, time))
-        self._value = numpy.hstack(
-            (self._value, value)
-        )  # so values must be numpy objects after all ?
-
-    def insert(self, time: int, value):
-        """modifies partition object to include value at time - other times are unchanged"""
-        assert not (time == self._time).any(), "this boundary exists already"
-        index = numpy.searchsorted(self._time, numpy.array([time]))[
-            0
-        ]  # move _this_ index to the right
-        self._time = numpy.hstack((self._time[:index], time, self._time[index:]))
-        # so values must be numpy objects after all ?
-        self._value = numpy.hstack((self._value[:index], value, self._value[index:]))
-
-    def delete_merge_left(self, index):
-        # TODO write unittests for me and my other half /
-        # or replace by just CUT if it's possible to uniquely specify the latter
-        """deletes a phoneme, leaves duration as is"""
-        assert len(self._value) > 1
-        assert 0 <= index < len(self._value)
-        self._time = numpy.hstack((self._time[:index], self._time[index + 1 :]))
-        self._value = numpy.hstack((self._value[:index], self._value[index + 1 :]))
-
-    def delete_merge_right(self, index):
-        """deletes a phoneme, leaves duration as is"""
-        assert len(self._value) > 1
-        assert 0 <= index < len(self._value)
-        self._time = numpy.hstack((self._time[: index + 1], self._time[index + 2 :]))
-        self._value = numpy.hstack((self._value[:index], self._value[index + 1 :]))
-
-    def merge_same(self):  # rename to uniq?
-        """return a copy of myself, except identical values will be merged"""
-        if len(self._value) <= 1:
-            return Partition(self._time, self._value, self.fs)
-        else:
-            time = [self._time[0]]
-            value = [self._value[0]]
-            for i, v in enumerate(self._value[1:]):
-                if v != value[-1]:  # new value
-                    time.append(self._time[i + 1])  # close last one, begin new one
-                    value.append(v)
-            time.append(self.duration)
-            return Partition(
-                numpy.array(time, dtype=TIME_TYPE), numpy.array(value), self.fs
-            )
-
-    def time_warp(self, X, Y):
-        assert X[0] == 0
-        assert Y[0] == 0
-        assert X[-1] == self.duration
-        time = numpy.interp(self.time, X, Y).astype(self.time.dtype)
-        # assert len(numpy.where(numpy.diff(time) <= 0)[0]) == 0,\
-        #     'some segment durations are non-positive' # Must allow this after all
-        # self._time = time
-        # may have to remove some collapsed items
-        self._time = time
-        while 1:
-            index = numpy.where(numpy.diff(self._time) == 0)[0]
-            if len(index):
-                logger.warning(
-                    'removing collapsed item #{}: "{}"'.format(
-                        index[0], self.value[index[0]]
-                    )
-                )
-                self.delete_merge_right(index[0])
-            else:
-                break
-
-
-class Event(Track):
-    def __init__(self, time, fs, duration, path=None):
-        super().__init__(path)
-        assert isinstance(time, numpy.ndarray)
-        assert time.ndim == 1
-        assert time.dtype == TIME_TYPE
-        assert (
-            numpy.diff(time.astype(numpy.float)) > 0
-        ).all(), "times must be strictly monotonically increasing"
-        assert isinstance(fs, int)
-        assert fs > 0
-        # assert isinstance(duration, TIME_TYPE) or isinstance(duration, int)
-        assert not (len(time) and duration <= time[-1]), "duration is not > times"
-        self._fs = fs
-        self._time = time
-        self._duration = TIME_TYPE(duration)
-
-    def get_time(self):
-        assert (
-            numpy.diff(self._time.astype(numpy.float)) > 0
-        ).all(), "times must be strictly monotonically increasing"
-        # in case the user messed with .time[index] directly
-        return self._time
-
-    def set_time(self, time):
-        assert isinstance(time, numpy.ndarray)
-        assert time.ndim == 1
-        assert time.dtype == TIME_TYPE
-        assert (
-            numpy.diff(time.astype(numpy.float)) > 0
-        ).all(), "times must be strictly monotonically increasing"
-        # assert (numpy.diff(time.astype(numpy.float)) >= 0).all(), "times must be strictly monotonically increasing"
-        assert not (len(time) and self._duration <= time[-1]), "duration is not > times"
-        self._time = time
-
-    def get_value(self):
-        raise Exception("No values exist for Events")
-
-    def set_value(self, value):
-        raise Exception("can't set values for Events")
-
-    def get_duration(self):
-        return self._duration
-
-    def set_duration(self, duration):
-        assert isinstance(duration, TIME_TYPE) or isinstance(duration, int)
-        assert not (
-            len(self._time) and duration <= self._time[-1]
-        ), "duration is not > times"
-        self._duration = duration
-
-    time = property(get_time, set_time)
-    value = property(get_value, set_value)
-    duration = property(get_duration, set_duration, doc="duration of track")
-
-    def __len__(self):
-        return len(self._time)
-
-    def __str__(self):
-        return "%s, fs=%i, duration=%i." % (self._time, self.fs, self.duration)
-
-    def __add__(self, other):
-        assert type(other) == type(self), "Cannot add Track objects of different types"
-        assert self.fs == other.fs, "sampling frequencies must match"
-        time = numpy.concatenate(
-            (self.time, (other.time + self.duration).astype(other.time.dtype))
-        )
-        duration = self.duration + other.duration
-        return type(self)(time, self.fs, duration)
-
-    def __eq__(self, other):
-        if (
-            (self._fs == other._fs)
-            and (self._duration == other._duration)
-            and (len(self._time) == len(other._time))
-            and (self._time == other._time).all()
-        ):
-            return True
-        else:
-            return False
-
-    def __ne__(self, other):
-        return not self.__eq__(other)
-
-    def crossfade(self, event, length):
-        """append wave to self, using a crossfade of a specified length in samples"""
-        assert type(self) == type(event), "Cannot add Track objects of different types"
-        assert self.fs == event.fs, "sampling frequency of waves must match"
-        assert isinstance(length, int)
-        assert length > 0
-        assert event.duration >= length
-        assert self.duration >= length
-        # cut left, cut right, concatenate
-        a = self.select(0, self.duration - length // 2)
-        b = event.select(length - length // 2, event.duration)
-        return a + b
-
-    def resample(self, fs):
-        if fs != self._fs:
-            factor = fs / self._fs
-            # need to use numpy.round for consistency - it's different from the built-in round
-            duration = int(numpy.ceil(factor * self._duration))
-            if len(self._time):
-                time = numpy.round(factor * self._time).astype(TIME_TYPE)
-                if (numpy.diff(time) == 0).any():
-                    logger.warning(
-                        "new fs causes times to fold onto themselves due to lack in precision, "
-                        "eliminating duplicates"
-                    )
-                    time = numpy.unique(time)
-                if duration <= time[-1]:  # try to fix this situation
-                    if len(time) > 1:
-                        if (
-                            time[-2] == time[-1] - 1
-                        ):  # is the penultimate point far enough away?
-                            raise Exception(
-                                "cannot adjust last time point to be smaller "
-                                "than the duration of the track"
-                            )
-                    logger.warning(
-                        "new fs causes last time point to be == duration, "
-                        "retarding last time point by one sample"
-                    )
-                    time[-1] -= 1
-            else:
-                time = self._time
-            return type(self)(time, fs, duration)
-        else:
-            return self
-
-    def select(self, a, b):
-        assert a >= 0
-        assert b > a
-        assert b <= self._duration
-        ai = self.time.searchsorted(a)
-        bi = self.time.searchsorted(b)
-        time = self._time[ai:bi] - a
-        return type(self)(time, self.fs, b - a)
-
-    default_suffix = ".evt"
-
-    @classmethod
-    def read(cls, name, fs, duration, *_args, **_kwargs):
-        """Loads object from name, adding default extension if missing."""
-        if name == os.path.splitext(name)[0]:
-            ext = cls.default_suffix
-            name += ext
-        else:
-            ext = os.path.splitext(name)[1].lower()
-        if ext == ".pml" or ext == cls.default_suffix:
-            self = cls.read_pml(name, fs)
-        elif ext == ".pp":
-            self = cls.read_PointProcess(name, fs)
-        else:
-            raise ValueError("file '{}' has unknown format".format(name))
-        if duration:
-            self.set_duration(duration)
-        return self
-
-    @classmethod
-    def read_pml(cls, name, fs=48000):
-        with open(name, "r") as f:
-            lines = f.readlines()
-        if len(lines) == 0:
-            logger.warning("pmlread(): empty file")
-            return Event(numpy.empty(0, dtype=TIME_TYPE), fs, 0)
-        time = numpy.zeros(len(lines), TIME_TYPE)
-        for i, line in enumerate(lines):
-            token = line.split(" ")
-            # t1 = token[0]
-            t2 = token[1]
-            time[i] = numpy.round(float(t2) * fs)
-        if (numpy.diff(time) <= 0).any():
-            logger.error(
-                "events are too close (for fs=%i) in file: %s, merging events"
-                % (fs, name)
-            )
-            time = numpy.unique(time)
-        # we cannot truly know the duration, so we are giving it the minimum duration
-        return Event(time, fs, int(time[-1] + 1))
-
-    pmlread = read_pml
-
-    @classmethod
-    def read_pm(cls, name, fs, _duration):
-        # suited for loading .pm files (pitch mark) that exist in CMU-ARCTIC
-        with open(name, "r") as f:
-            lines = f.readlines()
-        if len(lines) == 0:
-            logger.warning("pmread(): empty file")
-            return Event(numpy.empty(0, dtype=TIME_TYPE), fs, 0)
-        time = numpy.zeros(len(lines), TIME_TYPE) - 1
-        for i, line in enumerate(lines):
-            token = line.split(" ")
-            t1 = token[0]
-            with contextlib.suppress(IndexError):
-                time[i] = numpy.round(float(t1) * fs)
-            # try:
-            #     time[i] = numpy.round(float(t1) * fs)
-            # except IndexError:
-            #     continue
-            # else:
-            #     t2 = token[1]
-
-        time = time[time != -1]
-        if (numpy.diff(time) <= 0).any():
-            logger.error(
-                "events are too close (for fs=%i) in file: %s, merging events"
-                % (fs, name)
-            )
-            time = numpy.unique(time)
-        # if int(time[-1] + 1) >= duration:
-        # time = time[:-1]
-        # we cannot truly know the duration, so we are giving it the minimum duration
-        return Event(time, fs, int(time[-1] + 1))
-
-    def write(self, name, *_args, **_kwargs):
-        """Saves object to name, adding default extension if missing."""
-        name_wo_ext = os.path.splitext(name)[0]
-        if name == name_wo_ext:
-            name += self.default_suffix
-        self.write_pml(name)
-
-    def write_pml(self, name):
-        f = open(name, "w")
-        t1 = 0.
-        for t in self.time:
-            t2 = t / self.fs
-            f.write("%f %f .\n" % (t1, t2))
-            t1 = t2
-        f.close()
-
-    pmlwrite = write_pml
-
-    # def __getitem__(self, index):
-    #     return self._time[index]
-
-    # def __setitem__(self, index, value):
-    #     self._time[index] = value
-
-    def get(self, t):
-        if t in self._time:
-            return True
-        else:
-            return False
-
-    def draw_pg(self, **kwargs):
-        raise NotImplementedError
-
-    def time_warp(self, X, Y):
-        assert X[0] == 0
-        assert Y[0] == 0
-        assert X[-1] == self.duration
-        time = numpy.interp(self.time, X, Y).astype(self.time.dtype)
-        if 0:
-            # from matplotlib import pylab
-            # pylab.plot(X, Y, 'rx-')
-            # for x, y in zip(self.time, time):
-            #     pylab.plot([x, x, 0], [0, y, y])
-            # pylab.show()
-            raise NotImplementedError
-        # may have to remove some collapsed items
-        assert len(numpy.where(numpy.diff(time) == 0)[0]) == 0
-        self._time = time  # [index]
-        self.duration = Y[-1]  # changed this from _duration
-
-    #  TODO: NEEDS TESTING!
-    def insert(self, a, t):  #
-        assert isinstance(t, type(self))
-        index = numpy.where((self.time >= a))[0][0]  # first index of the "right side"
-        self._time = numpy.hstack(
-            (self._time[:index], t._time + a, self._time[index:] + t.duration)
-        )
-        self._duration += t.duration
 
 
 # class MultiTrack(dict):
@@ -2290,42 +441,6 @@ class Event(Track):
 #             )  # untested
 #         return obj
 
-
-# class TestEvent(unittest.TestCase):
-#     def setUp(self):
-#         self.t = Event(numpy.array([3, 6], dtype=TIME_TYPE), 1, 10)
-#         self.u = Event(numpy.array([0, 2], dtype=TIME_TYPE), 1, 3)
-
-#     def test_init(self):
-#         Event(numpy.array([], dtype=TIME_TYPE), 1, 0)  # empty
-#         Event(numpy.array([], dtype=TIME_TYPE), 1, 10)  # empty
-#         self.assertRaises(AssertionError, Event, numpy.array([6, 3], dtype=TIME_TYPE), 1, 10)  # bad times
-#         self.assertRaises(Exception, Event, numpy.array([3, 6], dtype=TIME_TYPE), 1, 5)  # duration too short
-
-#     def test_duration(self):
-#         self.t.duration = 8  # ok
-#         self.assertRaises(Exception, self.t.set_duration, 5)  # duration too short
-
-#     def test_eq(self):
-#         self.assertTrue(self.t == self.t)
-#         self.assertFalse(self.t == self.u)
-
-#     def test_add(self):
-#         v = self.t + self.u
-#         self.assertTrue(v == Event(numpy.array([3, 6, 10, 12], dtype=TIME_TYPE), 1, 13))
-#         self.t += self.u
-#         self.assertTrue(self.t == Event(numpy.array([3, 6, 10, 12], dtype=TIME_TYPE), 1, 13))
-
-#     def test_select(self):
-#         t = self.t.select(2, 7)
-#         self.assertTrue(t == Event(numpy.array([1, 4], dtype=TIME_TYPE), 1, 5))
-#         t = self.t.select(3, 7)
-#         self.assertTrue(t == Event(numpy.array([0, 3], dtype=TIME_TYPE), 1, 4))
-#         t = self.t.select(3, 6)
-#         self.assertTrue(t == Event(numpy.array([0], dtype=TIME_TYPE), 1, 3))
-#         t = self.t.select(2, 6)
-#         self.assertTrue(t == Event(numpy.array([1], dtype=TIME_TYPE), 1, 4))
-
 # def test_pml(self):
 #     import tempfile
 #     tmp = tempfile.NamedTemporaryFile(prefix='test_pml_')
@@ -2338,31 +453,6 @@ class Event(Track):
 #     s.duration = int(numpy.round(self.t.duration * s.fs / self.t.fs))
 #     s = s.resample(self.t.fs)
 #     self.assertTrue(numpy.allclose(s.time, self.t.time))
-
-
-# class TestWave(unittest.TestCase):
-#     def setUp(self):
-#         self.w = Wave(numpy.arange(0, 16000), 16000)
-#         self.v = Wave(numpy.arange(100, 200), 16000)
-
-# def test_init(self):
-#     t = Wave(numpy.array([], dtype=TIME_TYPE), 1)  # empty
-
-#     def test_eq(self):
-#         self.assertTrue(self.w == self.w)
-#         self.assertFalse(self.w == self.v)
-
-#     def test_add(self):
-#         t = self.w + self.v
-#         self.assertTrue(t.duration == 16100)
-#         self.assertTrue(t.value[16050] == 150)
-#         self.w += self.v
-#         self.assertTrue(self.w.duration == 16100)
-#         self.assertTrue(self.w.value[16050] == 150)
-
-#     def test_select(self):
-#         w = self.w.select(10, 20)
-#         self.assertTrue(w == Wave(numpy.arange(10, 20, dtype=TIME_TYPE), 16000))
 
 
 # class TestTimeValue(unittest.TestCase):
@@ -2432,69 +522,6 @@ class Event(Track):
 #         self.assertTrue(t == TimeValue(numpy.array([0, 4], dtype=TIME_TYPE), numpy.array([1, 4]), 1, 5))
 #         t = self.t1.select(2, 5)
 #         self.assertTrue(t == TimeValue(numpy.array([], dtype=TIME_TYPE), numpy.array([]), 1, 3))
-
-
-# class TestPartition(unittest.TestCase):
-#     def setUp(self):
-#         self.p1 = Partition(numpy.array([0, 5, 6, 10], dtype=TIME_TYPE), numpy.array(["start", "middle", "end"]), 1)
-#         self.p2 = Partition(numpy.array([0, 5, 10], dtype=TIME_TYPE), numpy.array(["start2", "end2"]), 1)
-
-#     def NOtest_init(self):  # should one allow an empty partition?
-#         Partition(numpy.array([0, 1], dtype=TIME_TYPE), numpy.zeros(1, dtype=TIME_TYPE), 1)
-#         # p = Partition(numpy.empty(0, dtype=TIME_TYPE), numpy.empty(0, dtype=TIME_TYPE), 1)
-
-#     def test_insert(self):
-#         p = Partition(numpy.array([0, 5, 6, 10], dtype=TIME_TYPE), numpy.array(["start", "middle", "end"]), 1)
-#         # p.insert(0, "even earlier")
-#         # p.insert(1, "really the end")
-#         p.insert(7, "still the middle")
-#         self.assertTrue((p.time == numpy.array([0, 5, 6, 7, 10])).all())
-
-#     def test_duration(self):
-#         self.p1.duration = 12  # ok
-#         self.p1.duration = 8  # ok
-#         self.assertRaises(Exception, self.p1.set_duration, 6)  # duration too short
-
-#     def test_eq(self):
-#         self.assertTrue(self.p1 == self.p1)
-#         self.assertFalse(self.p1 == self.p2)
-
-#     def test_add(self):
-#         p = self.p1 + self.p2
-#         self.assertTrue(p.duration == 20)
-#         self.assertTrue(p.time[4] == 15)
-#         self.assertTrue(p.value[4] == "end2")
-#         self.p1 += self.p2
-#         p = self.p1
-#         self.assertTrue(p.duration == 20)
-#         self.assertTrue(p.time[4] == 15)
-#         self.assertTrue(p.value[4] == "end2")
-
-#     def test_select(self):
-#         p = self.p1.select(5, 6)
-#         self.assertTrue(p == Partition(numpy.array([0, 1], dtype=TIME_TYPE), numpy.array(['middle']), 1))
-#         p = self.p1.select(5, 7)
-#         self.assertTrue(p == Partition(numpy.array([0, 1, 2], dtype=TIME_TYPE), numpy.array(['middle', 'end']), 1))
-#         p = self.p1.select(4, 6)
-#         self.assertTrue(p == Partition(numpy.array([0, 1, 2], dtype=TIME_TYPE), numpy.array(['start', 'middle']), 1))
-#         p = self.p1.select(4, 7)
-#         self.assertTrue(p == Partition(numpy.array([0, 1, 2, 3], dtype=TIME_TYPE),
-#                                        numpy.array(['start', 'middle', 'end']),
-#                                        1))
-
-#     def test_from_TimeValue(self):
-#         tv = TimeValue(numpy.arange(9, dtype=TIME_TYPE) * 10 + 10,
-#                        numpy.array([0.0, 1.0, 1.0, 4.0, 4.0, 4.0, 8.0, 8.0, 8.0]),
-#                        1,
-#                        100)
-#         p = Partition.from_TimeValue(tv)
-#         self.assertTrue((p.time == numpy.array([0, 15, 35, 65, 100])).all())
-#         self.assertTrue((p.value == numpy.array([0.0, 1.0, 4.0, 8.0])).all())
-
-#     def test_merge_same(self):
-#         p = Partition(numpy.array([0, 3, 6, 10], dtype=TIME_TYPE), numpy.array(["1", "1", "2"]), 1)
-#         p = p.merge_same()
-#         self.assertTrue(p.value[1] == "2")
 
 
 # class TestMultiTrack(unittest.TestCase):
